Channel specific default facts
If a channel is defined in a property file the facts from that file will
be incorporated into the defaults of the channel. This effectively creates a
cohort group for defaults. So for example all properties marked as social will
have an average emmmision per employee that is inclusive of social properties.